backend/requirements.txt
# Core dependencies
langchain==0.1.0
langchain-openai==0.0.2
langchain-community>=0.0.20
fastapi==0.109.0
uvicorn==0.25.0
python-dotenv==1.0.0
pydantic>=2.5.3
python-multipart==0.0.6

# Vector stores and embeddings
faiss-cpu==1.7.4
tiktoken==0.5.2

# Text processing
beautifulsoup4==4.12.2
pypdf==3.17.4

# Development
pytest==7.4.4
black==23.12.1
